Fixed component testing
