+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "prepare": "husky install",
-    "test": "vitest tests/components/*.test.ts",
+    "test": "vitest tests/**.test.ts",
     "test:ui": "vitest --ui tests/components/*.test.ts",
     "dev": "vite",
     "build": "vue-tsc && vite build",
